switched xy_sample_random to sf
Conversion of sf to owin leads to a different tile order and subsequently to different sampling points because of a different state of the RNG. Thus, I used a detour over SpatialPolygons to owin to get identical results (make the unit tests happy) for now. This should be cut out later, making maptools obsolete.